+++ .gitignore
@@ -1,6 +1,8 @@
 /target
 **/*.rs.bk
 
+Cargo.lock
+
 # for IDE
 .idea/
 .vscode/
