+++ Cargo.toml
@@ -1,14 +1,3 @@
-[package]
-name = "rust-ipfs"
-version = "0.1.0"
-authors = ["Aten <jincxmain@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-
-
 [workspace]
 members = [
     "core/multibase",
