+++ core/block-format/src/error.rs
@@ -6,7 +6,6 @@ use rust_cid::{Cid, Error as CidErr};
 pub enum BlockFormatError {
     #[error("data did not match given hash, fst: {0}, snd: {1}")]
     WrongHash(Cid, Cid),
-
     #[error("Cid Error {0}")]
     CidError(CidErr),
 }
