+++ core/cid/Cargo.toml
@@ -4,12 +4,9 @@ version = "0.1.0"
 authors = ["Aten <jincxmain@gmail.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-thiserror = "1.0.6"
-integer-encoding = "1.0.7"
+thiserror = "1.0"
+integer-encoding = "1.0"
 
-multihash = { package = "parity-multihash", git = "https://github.com/libp2p/rust-libp2p" }
-
+multihash = { package = "parity-multihash", version = "0.2" }
 multibase = { package = "rust-multibase", path = "../multibase" }
