+++ core/cid/src/error.rs
@@ -1,8 +1,7 @@
 use std::io;
-use thiserror::Error;
 
-use multibase;
 use multihash::Hash as MHashEnum;
+use thiserror::Error;
 
 pub type Result<T> = ::std::result::Result<T, Error>;
 
@@ -11,10 +10,8 @@ pub type Result<T> = ::std::result::Result<T, Error>;
 pub enum Error {
     #[error("Invalid hash bytes for cidv0, code:{:?}, digest len:{}", .0, .1)]
     InvalidCidV0(MHashEnum, usize),
-
     #[error("Invalid v0 prefix")]
     InvalidV0Prefix,
-
     #[error("Unknown codec")]
     UnknownCodec,
     #[error("Input too short")]
