+++ core/multibase/Cargo.toml
@@ -7,15 +7,13 @@ authors = ["Friedel Ziegelmayer <dignifiedquire@gmail.com>"]
 edition = "2018"
 
 keywords = ["ipld", "ipfs", "multihash", "multibase", "cid"]
-
 license = "MIT"
-
 readme = "README.md"
 
 [dependencies]
 base-x = "0.2"
 base32 = "0.4"
-base64 = "0.10"
+base64 = "0.11"
 
 [dev-dependencies]
 criterion = "0.3"
