+++ core/primitives/Cargo.toml
@@ -4,8 +4,4 @@ version = "0.1.0"
 authors = ["Aten <jincxmain@gmail.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-serde = { version = "1.0.101", features = ["derive"] }
-impl-serde = { version = "0.2.3" }
