+++ core/util/Cargo.toml
@@ -4,7 +4,5 @@ version = "0.1.0"
 authors = ["Aten <jincxmain@gmail.com>"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-multihash = { package = "parity-multihash", git = "https://github.com/libp2p/rust-libp2p" }
+multihash = { package = "parity-multihash", version = "0.2" }
